requirements.txt
colour==0.1.5
easydict==1.9
editdistance==0.6.0
imgaug==0.4.0
ipython==8.6.0
lmdb==1.3.0
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.5.64
Pillow==9.3.0
pygame==2.1.2
PyYAML==6.0
scipy==1.5.4
six==1.16.0
timm==0.6.5
torch==1.10.1+cu113
TorchSnooper==0.8
torchvision==0.11.2+cu113
tqdm==4.64.0
